lab/project/requirements.txt

black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cowsay==5.0
idna==3.4
mypy-extensions==0.4.3
pathspec==0.10.2
platformdirs==2.5.4
pyfiglet==0.8.post1
python-dotenv==0.21.0
python-nmap==0.7.1
requests==2.28.1
tomli==2.0.1
urllib3==1.26.13
